Refactor item updaters and update GildedRose class
This commit refactors the item updaters in the GildedRose project. The AgedBrieUpdater and BackstagePassUpdater classes have been modified to improve the quality update logic. The increaseQualityIfPossible and increaseQualityAfterSellInIfPossible methods have been renamed to increaseQuality and increaseQualityAfterSellIn, respectively. The ConjuredUpdater class has also been modified to use the decreaseQuality method instead of directly subtracting from the quality. Additionally, the GildedRose class has been updated to use the ItemUpdaterFactory to create the appropriate item updater based on the item name. This improves the extensibility of the code and allows for easier addition of new item types.